python script to generate email of weekly events
