+++ develop.sh
@@ -2,4 +2,4 @@
 mongod &
 virtualenv --no-site-packages .
 source bin/activate
-pip install -r config/requirements.txt
+pip install -r config/requirements.txt --allow-external PIL --allow-unverified PIL
