+++ docs/quickstart.rst
@@ -47,7 +47,7 @@ steps:
 
 4. Authorize the Google Calendar API::
 
-       $ python manage.py --authorize
+       $ python authorize.py
 
 Then you should be all set to run!
 
@@ -149,7 +149,8 @@ Organization / Structure
     ├── config           # Configuration files
     ├── data             # Backup data
     ├── manage.py        # Various scripts.  Run `python manage.py` to view usage.
+    ├── authorize.py     # Authorization script.  Run `python authorize.py` to authorize the Google Calendar API Client.
     ├── run.py           # Runs the app!
-    ├── script           # Scripts run by `manage.py` outside of the app
+    ├── script           # Scripts run by `manage.py` and 'authorize.py' outside of the app
     └── test             # Unit tests
